+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Implemented HTTP utils and API clients.
+- Implemented a SendGridClient.
 
 ## [1.11.0] - 2024-01-10